+++ src/lib.rs
@@ -54,7 +54,7 @@ mod module_writer;
 mod new_project;
 mod polyfill;
 mod project_layout;
-mod pyproject_toml;
+pub mod pyproject_toml;
 mod python_interpreter;
 mod source_distribution;
 mod target;