+++ test-crates/pyo3-mixed-include-exclude/check_installed/check_installed.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+
+import pyo3_mixed_include_exclude
+
+assert pyo3_mixed_include_exclude.get_42() == 42
+
+print("SUCCESS")