+++ test-crates/pyo3-mixed-include-exclude/pyo3_mixed_include_exclude/__init__.py
@@ -1,0 +1,6 @@
+from .python_module.double import double
+from .pyo3_mixed_include_exclude import get_21
+
+
+def get_42() -> int:
+    return double(get_21)